requirements.txt
smolagents[litellm]
python-dotenv
pytz
pyyaml
gradio
huggingface_hub
streamlit
ddgs
requests
duckduckgo-search
google-search-results
Pillow
litellm>=1.61.0
google-generativeai>=0.8.3
langchain-community>=0.3.0
